Extract EFL test helper
Move common methods for setting up an EFL-related system spec into a
module, and include it in the IELTS/TOEFL specs.
