+++ spec/system/candidate_interface/english_as_a_foreign_language/add_ielts_qualification_spec.rb
@@ -2,6 +2,7 @@ require 'rails_helper'
 
 RSpec.feature 'Add IELTS qualification' do
   include CandidateHelper
+  include EFLHelper
 
   scenario 'Candidate completes EFL section with details of their IELTS' do
     given_i_am_signed_in
@@ -18,26 +19,6 @@ RSpec.feature 'Add IELTS qualification' do
 
   def given_i_am_signed_in
     create_and_sign_in_candidate
-  end
-
-  def and_the_efl_feature_flag_is_active
-    FeatureFlag.activate(:efl_section)
-  end
-
-  def and_i_declare_a_non_english_speaking_nationality
-    visit candidate_interface_application_form_path
-    click_link 'Personal details'
-    candidate_fills_in_personal_details
-    click_link 'Personal details'
-    click_link 'Change nationality'
-    select 'Hong Konger', from: 'Nationality'
-    select 'Pakistani', from: 'Second nationality'
-    click_button 'Save and continue'
-    click_button 'Continue'
-  end
-
-  def and_i_click_on_the_efl_section_link
-    click_link efl_link_text
   end
 
   def and_i_select_the_options_for_ielts
@@ -82,11 +63,5 @@ RSpec.feature 'Add IELTS qualification' do
     check 'I have completed this section'
     click_button 'Continue'
     expect(page).to have_css('#english-as-a-foreign-language-badge-id', text: 'Completed')
-  end
-
-private
-
-  def efl_link_text
-    'English as a foreign language'
   end
 end
